AAAAAAAAAAAAAAAAAAAAAAAA

+++ main.py
@@ -7,7 +7,7 @@ from matplotlib.figure import Figure
 from PyQt5.QtCore import Qt, QDate
 from PyQt5.QtGui import QFont
 from PyQt5.QtWidgets import QApplication, QMainWindow, QWidget, QTableWidgetItem, QTreeWidgetItem, QPushButton, \
-    QComboBox, QTableWidget, QSizePolicy, QDateEdit
+    QComboBox, QTableWidget, QSizePolicy, QDateEdit, QLabel
 from PyQt5 import uic, QtWidgets
 
 user = None
@@ -131,7 +131,11 @@ class Task(QWidget):
         super().__init__()
         uic.loadUi('tasks.ui', self)
         self.setMouseTracking(True)
-        self.pb_addT.clicked.connect(self.addTask)
+        self.role = cur.execute(f'''SELECT admin FROM main WHERE SN="{user}"''').fetchall()[0][0]
+        if self.role == 'Admin':
+            self.pb_addT.clicked.connect(self.addTask)
+        else:
+            self.pb_addT.setParent(None)
         self.pb_reboot.clicked.connect(self.reboot)
         self.name = name
         self.poz = -1
@@ -142,8 +146,12 @@ class Task(QWidget):
         self.rowNum = None
         self.pb_more = None
         self.ispolniteli = []  # Переменная хранящая всех сотрудников
-        self.status = ['', 'Удалить', 'Обновить']
-        self.status.extend(rowTitles)
+        self.rT = rowTitles[:]
+        if self.role == 'Admin':
+            self.rT.extend(['Удалить', 'Обновить'])
+        elif self.role == 'Edit':
+            self.rT.extend(['Обновить'])
+        self.status = self.rT
         # В двумерных списках помещены параметры задач, например self.cbs[Номер вкладки][Номер задачи](с нуля)
         self.cbs = [[] for _ in range(len(rowTitles))]  # combobox с исполнителями
         for i in range(len(cur.execute('''SELECT id FROM main''').fetchall())):
@@ -185,20 +193,22 @@ class Task(QWidget):
                     self.tabs[self.c_num].insertRow(0)
                 else:
                     self.tabs[self.c_num].setRowCount(1)
-                self.cbs[self.c_num].append(QComboBox())
-                self.cbs[self.c_num][-1].addItems(self.ispolniteli)
-                try:
+                if self.role in ['Admin', 'Edit']:
+                    self.cbs[self.c_num].append(QComboBox())
+                    self.cbs[self.c_num][-1].addItems(self.ispolniteli)
                     self.cbs[self.c_num][-1].setCurrentIndex(self.ispolniteli.index(self.sn))
                     self.sn = None
-                except:
-                    pass
+                else:
+                    self.cbs[self.c_num].append(QLabel(self.sn))
+                    self.sn = None
                 self.dts[self.c_num].append(QDateEdit())
                 self.dtss[self.c_num].append(QDateEdit())
-                try:
-                    self.dts[self.c_num][self.rowNum].setDate(self.startdate1)
-                    self.dtss[self.c_num][self.rowNum].setDate(self.enddate1)
-                except:
-                    pass
+                self.dts[self.c_num][self.rowNum].setDate(self.startdate1)  # Работать ЗДЕСЬ
+                self.dtss[self.c_num][self.rowNum].setDate(self.enddate1)
+
+                if self.role == 'False':
+                    self.dts[self.c_num][self.rowNum].setReadOnly(True)
+                    self.dtss[self.c_num][self.rowNum].setReadOnly(True)
                 self.pb_more = QPushButton('Подробнее')
                 self.pbs[self.c_num].append(self.pb_more)
                 self.cbss[self.c_num].append(QComboBox())
@@ -319,13 +329,20 @@ class Kanbaner(QMainWindow):
     def __init__(self):
         super().__init__()
         uic.loadUi('main.ui', self)
+        self.role = cur.execute(f'''SELECT admin FROM main WHERE SN="{user}"''').fetchall()[0][0]
         self.label.setText(user)
-        self.pb_create.clicked.connect(self.creater)
+        if self.role == 'Admin':
+            self.pb_create.clicked.connect(self.creater)
+            self.pb_delete.clicked.connect(self.delete)
+            self.pb_finance.clicked.connect(self.cash)
+            self.pb_graph.clicked.connect(self.graphics)
+        else:
+            self.pb_create.setParent(None)
+            self.pb_finance.setParent(None)
+            self.pb_delete.setParent(None)
+            self.pb_graph.setParent(None)
         self.pb_open.clicked.connect(self.open)
-        self.pb_delete.clicked.connect(self.delete)
         self.pb_login.clicked.connect(self.exit)
-        self.pb_finance.clicked.connect(self.cash)
-        self.pb_graph.clicked.connect(self.graphics)
         self.title = ''
         self.rowTitlesBad = []
         self.rowTitles = []
@@ -405,15 +422,15 @@ class Kanbaner(QMainWindow):
     def keyPressEvent(self, event):
         if event.key() == 16777220:
             self.open()
-        elif event.key() == Qt.Key_Backspace:
+        elif event.key() == Qt.Key_Backspace and self.role == 'Admin':
             self.delete()
-        elif event.key() == Qt.Key_Delete:
+        elif event.key() == Qt.Key_Delete and self.role == 'Admin':
             self.delete()
-        elif event.key() == Qt.Key_Q:
+        elif event.key() == Qt.Key_Q and self.role == 'Admin':
             self.creater()
-        elif event.key() == Qt.Key_W:
+        elif event.key() == Qt.Key_W and self.role == 'Admin':
             self.graphics()
-        elif event.key() == Qt.Key_E:
+        elif event.key() == Qt.Key_E and self.role == 'Admin':
             self.cash()
 
     def exit(self):
